allow the value of move entry to be a function

+++ src/move-files.js
@@ -2,21 +2,30 @@ import minimatch from 'minimatch'
 
 export default function (move) {
   return ctx => {
-    if (!move) return
-
     if (typeof move === 'function') {
       move = move(ctx.meta.merged)
     }
 
+    if (!move) return
+
     for (const pattern in move) {
       const matches = minimatch.match(ctx.fileList, pattern)
       if (matches.length > 0) {
-        const file = ctx.file(matches[0])
-        for (const match of matches) {
-          ctx.deleteFile(match)
-        }
         const newName = move[pattern]
-        ctx.createFile(newName, file)
+        if (typeof newName === 'function') {
+          for (const match of matches) {
+            const file = ctx.file(match)
+            const fileName = newName(match)
+            if (fileName) {
+              ctx.deleteFile(match)
+              ctx.createFile(fileName, file)
+            }
+          }
+        } else if (typeof newName === 'string') {
+          const file = ctx.file(matches[0])
+          ctx.deleteFile(matches[0])
+          ctx.createFile(newName, file)
+        }
       }
     }
   }
